store/wizard: move cloud provider selectors to cloud module
Keep selectors co-located with the state they access instead of
bundling them in the top-level wizard slice.
