+++ src/store/slices/wizard/cloud/index.ts
@@ -1,2 +1,3 @@
+export * from './selectors';
 export { initialState as cloudProvidersState } from './state';
 export * from './types';
